# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "Unisurvey"
  description: UniSurvey is an utility web service with the focus on improving students life with surveys and in-campus functionalities.
  license:
    name: MIT
    url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
host: polls.apiblueprint.org
basePath: /api/
schemes:
- https
consumes:
- application/json
produces:
- application/json
paths:
...
  /students/login:
    post:
       tags: [Student]
       summary: log into an existing student
       description: the student is logged in after validation of his email and password. A JWT token is returned.
           
       responses:
           '400':
               description: 'Bad request'
               
           '404':
               description: 'email not found'
               
           '401':
               description: 'wrong password: failed authentication'
           '200':
               description: 'user logged in'
...
  /admins/login:
    post:
        tags: [Admin]
        summary: log into an existing admin
        description: the admin is logged in after validation of his email and password. A JWT token is returned.
        responses:
            '400':
                description: 'Bad request'

            '404':
                description: 'email not found'

            '401':
                description: 'wrong password: failed authentication'

  
            '500':
                description: 'database internal error'

            '200':
                description: 'user logged in'
...
  /events:
    get:
        tags: [Events]
        summary: search all events
        description: a list of all events created is returned.
        responses:
            '500':
                description: 'database internal error'

            '200':
                description: 'Success: return the list of evetns'
...